feat: :art: incorporated new structure
\
