+++ packages/account/src/Sections/Security/ApiToken/api-token.scss
